feat: #17 패키지 이름 변경 및 Meet 관련 DTO 수정
DTO에서 반복적으로 구현되는 메소드에 대해 주석을 삭제했다.

MeetDTO에서 불필요하게 MeetAddressRepository Autowired 하는 부분이 있어서 이 부분을 삭제했다.

DTO에 Builder를 없애고 AllArgsConstructor를 private 접근으로 변경했다.
